# docs/api/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "网格社区治理平台 API",
    "description": "网格社区治理平台后端接口文档 v1.2.0 — 完整覆盖所有控制器",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/work-orders": {
      "get": {
        "tags": [
          "工单-Web"
        ],
        "summary": "分页查询工单列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "areaId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/work-orders/export": {
      "get": {
        "tags": [
          "工单-Web"
        ],
        "summary": "导出全量工单",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/work-orders/{id}": {
      "get": {
        "tags": [
          "工单-Web"
        ],
        "summary": "工单详情",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/work-orders/{id}/handle": {
      "post": {
        "tags": [
          "工单-Web"
        ],
        "summary": "处理工单节点",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/work-orders/{eventId}/dispatch": {
      "post": {
        "tags": [
          "工单-Web"
        ],
        "summary": "派发工单",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/work-orders/{id}/attachments": {
      "get": {
        "tags": [
          "工单-Web"
        ],
        "summary": "工单附件列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/work-orders/{id}/confirm-close": {
      "post": {
        "tags": [
          "工单-Web"
        ],
        "summary": "确认关闭工单",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/work-orders/{id}/reject-close": {
      "post": {
        "tags": [
          "工单-Web"
        ],
        "summary": "驳回关闭",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/work-orders/batch-delete": {
      "post": {
        "tags": [
          "工单-Web"
        ],
        "summary": "批量删除工单",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/h5/work-orders/{id}": {
      "get": {
        "tags": [
          "工单-H5"
        ],
        "summary": "工单详情",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/h5/work-orders/{id}/handle": {
      "post": {
        "tags": [
          "工单-H5"
        ],
        "summary": "处理工单",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/safety/inspections": {
      "get": {
        "tags": [
          "安全检查"
        ],
        "summary": "安全检查列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "安全检查"
        ],
        "summary": "新增检查记录",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/safety/inspections/statistics": {
      "get": {
        "tags": [
          "安全检查"
        ],
        "summary": "检查统计",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/safety/inspections/{id}/status": {
      "put": {
        "tags": [
          "安全检查"
        ],
        "summary": "更新整改状态",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/safety/inspections/mark-overdue": {
      "post": {
        "tags": [
          "安全检查"
        ],
        "summary": "标记超期未整改",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}